`lsp-imenu`: support categorizing by :kind? (#2313)
* `lsp-imenu': support categorizing by :kind?

* `lsp-imenu': use `lsp-symbol-kinds'

Rename `lsp--symbol-kind' to `lsp-symbol-kinds', making the former an
alias for the latter and the latter a `defcustom'. In it, rename "Enum
Member" to "Enumerator" since that looks better.

Instead of creating another constant for that does the same as
`lsp-symbol-kinds', use the latter.

Refactor: simplify `lsp--get-symbol-type' to use `alist-get' instead of
a complex threading sequence, simplifying it to a single s-expression.

* `lsp-imenu': use a separate SymbolKinds alist

* PR review: nit: use `lsp:document-symbol-kind'

* Terms: use "Enum Member" instead of "Enumerator"

The latter is specific to C/C++, and not used by the Java language
specification, for example. Use "Enum Member" instead, as it is used in
the LSP-specification

* Docstring: spelling: remove double "the"

* `lsp-imenu-index-function': fix customization

Remove extraneous sharp quotes, causing errors when using one of the
selected customization entries.

* `lsp-imenu-create-uncategorized-index': docstring

Clarify that the index created corresponds to the server's result directly.

* Add a new categorized `imenu' index

`lsp-imenu-create-categorized-index' becomes
`lsp-imenu-create-top-level-categorized-index', with the former now now
creating a "traditional" `imenu' index where all Functions are under a
`Functions' section, all Variables under `Variables', etc. It works with
both `SymbolInformation' and `DocumentSymbol's (tested with `pyright'
and `iph', respectively).

Currently there is no way to jump to a symbol if it has children; this
defect is shared by `org-imenu', because `imenu' entries with children
cannot encode a position.

* Include symbols with children in `imenu'

* Handle unknown symbol types: use ? as type

* Fix defcustom `const': point to right function

* `lsp-imenu': `defcustom': fix old function

`lsp-imenu-create-top-levelcategorized-index' was renamed to
`lsp-imenu-create-top-level-categorized-index'.