`lsp-install': use `compile'
Refactor `lsp-install-server' to use `compile' instead of
`make-process', with `compilation-shell-minor-mode'. The main benefit is
that ansi escape sequences, as used to display spinners and colours, now
work. Separate buffers are also tidier.

The `lsp--download-status' button had to be refactored to present a list
of installation buffers if clicked, now with a separate command,
`lsp-select-installation-buffer', instead of a lambda bound. For this,
`lsp-async-start-process' sets the `lsp--installation-buffer' flag in
new installation buffers, which is used to filter out installation
buffers from `buffer-list'.