Use compilation for `lsp-install` buffers (`lsp-async-start-process`) (#2677)
* `lsp-install': use `compile'

Refactor `lsp-install-server' to use `compile' instead of
`make-process', with `compilation-shell-minor-mode'. The main benefit is
that ansi escape sequences, as used to display spinners and colours, now
work. Separate buffers are also tidier.

The `lsp--download-status' button had to be refactored to present a list
of installation buffers if clicked, now with a separate command,
`lsp-select-installation-buffer', instead of a lambda bound. For this,
`lsp-async-start-process' sets the `lsp--installation-buffer' flag in
new installation buffers, which is used to filter out installation
buffers from `buffer-list'.

* `lsp-select-installation-buffer': C-u -> show dead

With a prefix C-u, `lsp-select-installation-buffer' now also shows
finished installation buffers.

* Mark finished/running installation buffers

`lsp-select-installation-buffer' now `propertize's installation buffers
with either `lsp-installation-buffer-face' or
`lsp-installation-finished-buffer-face' depending on whether the
installation is still in progress for that buffer or not.

* `lsp-async-start-process': fix bugs

- `combine-and-quote-strings' doesn't work with some characters, e.g
  backquotes; use `mapconcat' with `shell-quote-argument' instead
- After compilation finishes, the buffer has no process associated with
  it, causing a nil-error in case installation fails in `lsp-async-start'

* Add a command to delete stale installation buffers

Refactor the *lsp-installation*-buffer check into a new function,
because that way `lsp-async-start-process' could be made to use a
minor-mode in the future more easily.

* `lsp-async-start-process': use a minor-mode

This will allow users to set-up custom keybindings and to detach an
installation buffer.