Make sure that unchanged goes back to original buffer
Fixes #2701

there are cases in which post-command hook does not run in order to cancel the
request in 'unchanged mode, thus make sure that we are in the original buffer
when running the callback. Note that canceling the request is not critical and
we can afford those calls to run without canceling them.