Fix handling of lsp-file-watch-ignored-[directories|files] (#2736)
When either lsp-file-watch-ignored-directories or
lsp-file-watch-ignored-files are buffer local, the correct value may
not be used when a new directory is added to an already watched
directory. This is due to the watch callback invoking at any time. A
different buffer could be the active buffer when the callback is
invoked causing a different value of either of those two variables to
be used.

This was addressed by only supporting customization of these variables
at either the global level or at the root of an lsp workspace. When
creating a watches for a root lsp-mode now captures the values of the
lsp-file-watch-ignored-[directories|files] and passes them along to be
used throughout. This is done by creating a temporary buffer,
assigning it a name underneath the root, and hacking the local
variables in. This method was chosen as it minimizes the need of
hitting the disk to get the values of the variables. It fully supports
.dir-locals.el, specified either directly via a file, or via the class
mechanism describe in the Emacs documentation.

The signature of the following functions to changed to permit passing
the captured ignored-files and ignored-directories variables:
* lsp-watch-root-folder
* lsp--directory-files-recursively
* lsp--folder-watch-callback