Handle 'renamed' file notify event
lsp-mode only handles 'created', 'deleted' and 'changed' file watch events. At least on Windows 10, file watchers as created by file-notify-add-watch from filenotify.el raise a 'renamed' event. Since this event is currently ignored by lsp-mode, the lsp server is not notified that the file was renamed, and this typically lead the server into believing that both the original file and new file exist concurrently, which in itself lead to spurious error reports, double or ambiguous declarations, etc, depending on the language itself.

This commit handles properly the 'rename' event by by notifying the lsp server of the deletion of the file and creation of the file with the new name.