Use run-with-idle-timer for lsp-deferred (#3034)
`run-with-timer` triggers immediately after the buffer is visible. But
the environment is not set for the remote files at the time when LSP
binary is looked up.  This results in `matching-clients` to be nil.

Changing it to `run-with-idle-timer` ensures the buffer is done with
loading remote file and setting the environment and runs immediately
after the file is loaded in the buffer.