When auto-configure is enabled, remove hooks for disabled options
When `lsp-auto-configure' is enabled, hooks are automatically added to
`lsp-configure-hook' for any options which are enabled.  However, if any
of those options are later disabled, the hooks were not removed from
`lsp-configure-hook`; therefore the corresponding behaviour did not
get disabled.

So add the relevant `remove-hook' calls when options are not enabled.

Partial fix for #2525.