Call lsp--auto-configure from lsp-configure-buffer
Currently `lsp' is the only thing calling `lsp--auto-configure', which
means that auto-configuration only activates when `lsp-mode' called,
rather than when a buffer needs to be configured.

If we consider the following flow:

1. An option `lsp-headerline-breadcrumb-enable' (say) is set to `t'.
2. `lsp-mode` is called, triggering `lsp--auto-configure', which in
   turn executes:

     (add-hook 'lsp-configure-hook 'lsp-headerline-breadcrumb-mode)
3. `lsp-headerline-breadcrumb-enable' is set to `nil'.
4. Another buffer is opened.
5. At this point, `lsp-headerline-breadcrumb-mode' is still in
   `lsp-configure-hook', therefore, the breadcrumb headerline will be
   activated even though the option is disabled.  Clearly this is
   undesirable, as reported in #2525.

So call `lsp--auto-configure` from `lsp-configure-buffer', so that
any hooks which need to be removed will get removed.

Fixes #2525.