Fix documentation editing
Correctly resolve all editable links on the documentation (blog posts
and changelogs are left as-is, this can be easily modified).

There is now an expectation to find a `root_file` which is a string of a
file to resolve from the root of the lsp-mode tree.

The way we discover which files belong where is by:
- First checking if we can evaluate `symbol-file` for the group that
matches the name of the client (eg "lsp-csharp")
- Then checking for a file name in clients for a common group name
- Lastly checking for a file in clients (eg "clients/lsp-csharp.el")

Fixes #3413