Begin to support the new rust-analyzer inlay hints
* Lifetime elision hints
* Reborrow hints
* Closure return type hints

The way rust-analyzer is evolving, the emacs-side formatting looks
like it will become less and less useful, as these new hint types (and
indeed the already-existing chaining hint type) are not distinguished
in a way that the mode can determine; right now, I've left the support
in, but I've disabled it by default in preference to letting the
server format them.  In that case, the server provides a small amount
of formatting info (in particular, whether the inlay should be
formatted with padding on the left and/or right).