lsp-rust: set UPDATE_EXPECT=1 when expect-test on runnable (#3597)
The rust-analyzer experimental/runnables LSP extension includes an optional flag expectTest

This is set when the cursor is on an expect! or expect_file! macro when the lsp-rust-analyzer-run command is invoked.

Update the lsp-rust client to include env UPDATE_EXPECT=1 before the corgo command when this is set, so expect automatically updates the expected result to match actual.