Three modes, default, space and "newline"
The third mode is only reachable if you use the function on your
own (I use it to kill the returned type since this is a feature that
exists in the previous OCaml mode I used)

Thanks to @yyoncho I moved my code to the lsp-ocaml.el file to make it
OCaml related. The ocaml-lsp server returns a markdown file with the
following content:

```ocaml
<type>
---
<doc>
```

so it's easy to split it to only keep the signature.

If the second mode is chosen the type will be edited to be displayed
on one line and end with `...` if it is too wide