Don't prompt for an action within `lsp-organize-imports`, even if `lsp-auto-execute-action` is disabled.
Prior to this change invoking `lsp-organize-imports` would prompt for a choice
of action, if the variable `lsp-auto-execute-action` were set to nil. This seems
unintentional.

`lsp-auto-execute-action` is useful when `lsp-execute-code-action` is invoked
directly and the user doesn't necessarily know what action would be executed
automatically. But for a function like `lsp-organize-imports` created via
`lsp-make-interactive-code-action`, we know the user asked for a specific action
to be performed, so it doesn't make sense to prompt them again.

I encountered this behavior because I've set up the before-save hooks
[officially recommended
here](https://github.com/golang/tools/blob/master/gopls/doc/emacs.md#loading-lsp-mode-in-emacs)
for the Go language; with the `lsp-organize-imports` hook and `(setq
lsp-auto-execute-action nil)`, saving prompts every time.