lsp-rust: add semantic token modifiers
This is more of a proof-of-concept than anything else.

It works for the "mutable" modifier, adding an underline. This matches
what the VsCode client does, but clashes with the lsp-mode default for
read vs write highlighted variables.

The "reference" modifier has the bold attribute set.

The rest have a no-operation face assigned, but need to be there to
get the font caluclation numbers right, and to be customized by users.