Upgrade lsp-xml version and add new formatting options
This upgrades the language server version to 0.21.0 and adds all new
formatting options available. The default values are the defaults used
in the source code.

This also replaces obsolete `lsp-xml-format-quotations` with
`lsp-xml-preferences-quote-style`. Since they're not
compatible (instead of "doubleQuotes" and "singleQuotes" the values
are "double" and "single"), it was not marked as obsolete.

This also changes (or fixes?) `lsp-xml-validation-schema` being
defined as a boolean (when it's expected to be an object) to an alist.