Fix usage of ...trigger-for-incomplete-completions
Ensure that the completion session is the same as the previous
incomplete completion before using trigger style
`lsp/completion-trigger-kind-trigger-for-incomplete-completions`.

Previously the internal function `lsp-completion--get-context` had no
notion of whether the current completion request was the using the
same context as the previous completion request. This internal
function is only called by `lsp-completion-at-point`.
`lsp-completion-at-point` does calculate whether this completion
request is using the same context as the last result. We therefore now
pass the information from `lsp-completion-at-point` down to
`lsp-completion--get-context` and use it therein when making a
determination about which trigger type of completion to use.

This addresses issue #3028 where completions would stop working with
the sourcekit-lsp language server for Swift. This language server in
particular sends back incomplete completions frequently and also
checks on its side to make sure any requested completion continuation
context is the same as it was when originally requested. If not it
cancels the request.