Allow to not cut the signature received on hover (#3637)
* Allow to not cut the signature received on hover

Some servers (i.e. the ocamllsp one) return only the
signature on hover (without the documentation) but tend to put newlines between
the different items. With the current implementation of
`lsp-clients-extract-signature-on-hover`, the displayed signature will
only contain the first item

```
config1:string list ->
config2:string list -> config3:string list -> int -> int -> int
```

will then be rendered as

```
config1:string list ->
```

With the new implementation, a custom variable allows to take control
over this behaviour and the function has been reimplemented to either
return the previous result or return the signature with newlines
replaced by spaces

* Three modes, default, space and "newline"

The third mode is only reachable if you use the function on your
own (I use it to kill the returned type since this is a feature that
exists in the previous OCaml mode I used)

Thanks to @yyoncho I moved my code to the lsp-ocaml.el file to make it
OCaml related. The ocaml-lsp server returns a markdown file with the
following content:

```ocaml
<type>
---
<doc>
```

so it's easy to split it to only keep the signature.

If the second mode is chosen the type will be edited to be displayed
on one line and end with `...` if it is too wide

* Small updates and better function documentation

* Update lsp-mode.el