Prevents CPU throttling issue when LS doesn't support semantic token
This fixes the issue that is easily reproducible with the haskell
language server. Steps to reproduce:

- Enable semantic tokens

(setq lsp-semantic-tokens-enable t)

- Configure haskell language server and open some file.
- You could observe that your CPU will be throttled and will have a
  high consumption.

Enabling the logs, you can see that the lsp client will be constantly
send semantic token requests to hls which will be contiously returing
-32601.

Notes regarding the changes:

- Right now (lsp-feature? "textDocument/semanticTokens") will always
return true no matter what. Infact, even an arbitrary value
like (lsp-feature? "some_value") returns true. So, I don't think it's
a correct check. And I believe this happens because of the missing
`:check-command` in the `lsp-method-requirements`.
- Changing it to "textDocument/semanticTokensFull" fixes the issue for
me. And I belive this is because it has an appropriate `:check-command`.