headerline: Fix error for files with symlinks
`lsp--workspace-root` (returned by `lsp-headerline--workspace-root`)
always stores a fully resolved path with no symlinks.

When `(buffer-file-name)` includes a symlinked path to the workspace
root, `root` is not a prefix of `(buffer-file-name)`.
In this case, the path parents traversing in
`lsp-headerline--path-up-to-project-root` doesn't stop at root
and ultimately fails with a `(wrong-type-argument stringp nil)` error
when "/" is reached.

Fix this by resolving symlinks in `(buffer-file-name)`.

Also simply use `file-name-directory` because `(buffer-file-name)` is
always an absolute path to a file.