Fix: binary download for linux/aarch64 (csharp and rust)
we use the lsp--system-arch (csharp) and lsp--gnu-arch (rust) constructs
to generate an url for the right version of the binary for any os/arch
variant.

if we ever get an windows-nt/sh machine, we'll at least try to get the
right binary for it, and probably fail because no-one ever will build
such abomination.

for csharp, this patch rewrites the logic using pcase as it's been used
in the rest of the codebase.