fix[headerline]: use lsp-workspace-root to avoid issues with multiple active workspaces
Commit 16b13adf857f9cd2e7d48a1bfe3c0b1ed4107339 introduced a new function lsp-headerline--workspace-root aimed at speeding up the headerline update function, as lsp-workspace-root is expensive (https://github.com/emacs-lsp/lsp-mode/issues/3911)

Unfortunally, the strategy assumed that the buffer-local lsp--buffer-workspace would only contain workspaces that the current buffer belongs to. This assumption appears to be wrong.

As a result, when two buffers belonging to different workspaces (/path/to/workspace1 and /path/to/workspace2) had the headerline active, the idle function on the workspace2 buffer could actually assume it's workspace root was /path/to/workspace1. 

As suggested by @yyoncho in https://github.com/emacs-lsp/lsp-mode/commit/16b13adf857f9cd2e7d48a1bfe3c0b1ed4107339#r98855365, we can cache the result of lsp-workspace-root and have the faster headerline introduced by @erikarvstedt and avoid the bug altogether.