Don't block within lsp-request-while-no-input
blocking seems to be at odds with the function name, and can
be annoying if the language server fails to respond to a
completion request (which is currently the only request
handled by `lsp-request-while-no-input`). To restore the
former behavior (of having lsp-request-while-no-input block
whenever non-essential is nil), set the newly introduced
defcustom `lsp-request-while-no-input-may-block` to t.