Improve documentation for lsp-execute-code-action
`lsp-execute-code-action` is a user-facing function that can be
executed e.g. when clangd says "fix available". But as it stands, the
function description is too abstract, so unless you know details
around lsp-protocol `codeAction`, you won't figure out what it's
for. Fix that by clarifying what a user can use it for.