Add configuration for lenses in lsp-rust.el:
Most of the lenses are enabled by default in rust-analyzer if the LSP
client doesn't disable them, and so happen to work, if only by
accident. However, the References lens is disabled by default and this
patch allows it to be enabled. It also allows the other lenses to be
disabled if desired.