Parse JSON false as :json-false instead of nil
This would break code actions where the server sends us an action
containing a `false` boolean parameter. `lsp-mode` would translate this
into `nil`, and then send a `null` back to the language server. An
example of this causing problems was
https://github.com/emacs-lsp/lsp-mode/issues/4184.

There may be parts of `lsp-mode` that expect `false` values to be parsed
as `nil`, so any logic involving parsed booleans may need to be updated.