ci: bump: GitHub Actions setup-cmake and setup-python
What
===
I Bump GitHub Actions setup-cmake and setup-python.

Why
===
This may be due to the use of old Actions, or because an unsupported version of Node.js is being used.
Numerous warnings are being generated.
Security support has expired, which may cause problems, even though it is only used for CI.
In any case, it is best to remove the warnings.