Provide better Ada semantic highlighting defaults. (#4321)
The Ada Language Server provides numerous semantic token modifiers
that can be used, but most of the modifiers should not affect the
original base token face.  This change provides a mechanism to
customize the faces associated with semantic tokens and modifiers that
are applied to the buffer, while also providing better defaults for
out-of-the-box usage.