Revert "lsp-csharp: add a test to check if dotnet tool csharp-ls is actually installed" (#4368)
* treewide: replace `looking-at` with `looking-at-p` if match data unused

`looking-at-p` does not change match data and so should be a bit faster.

* Revert "lsp-csharp: add a test to check if dotnet tool csharp-ls is actually installed"

This reverts commit 9b3a9215807af0727b514e8c7cf440bcc0bdad44.

This commit broke detection of `csharp-ls`. I am not sure what was the
purpose of the change. The function it adds says it checks whether
`csharp-ls` is present in the system, however the function
`lsp-csharp--cls-make-launch-cmd` always did the same and better. So
why have this additional check is not at all clear.

The function was assuming that a `dotnet tool list -g` should return
if `csharp-ls` is installed, however for me its output is just empty:

    λ dotnet tool list -g
    Package Id      Version      Commands
    -------------------------------------

Let's revert this commit.

Fixes: https://github.com/emacs-lsp/lsp-mode/issues/4365