lsp-mode.el: remove duplicated `nil` choice in lsp-document-sync-method
Worth noting that the two descriptions contradict each other. In the
code, the only place where the variable is handled is in
`lsp-on-change` function. Specifically, if the variable is nil, it
calls `(lsp--workspace-sync-method workspace)` which presumably
implies that the description "use method recommended by the
lang-server" is the correct one. So remove the other.

And while at it, remove the `lsp--sync-none` which was never used.

Fixes:
    lsp-mode.el:489:19: Warning: in defcustom for ‘lsp-document-sync-method’:
        duplicated value in ‘choice’: ‘nil’