Fix a number of warnings (#4373)
* lsp-javascript.el: fix incorrect default value in defcustom choices

Fixes a warning:

      lsp-javascript.el:488:12: Warning: in defcustom for ‘lsp-typescript-locale’: ‘nil’ is not a valid type

* lsp-mode.el: don't use let* for a single binding

`let*` vs `let` are only different in the order of binding the
variables, however it doesn't matter when there's just one variable.

* lsp-mode.el: remove duplicated `nil` choice in lsp-document-sync-method

Worth noting that the two descriptions contradict each other. In the
code, the only place where the variable is handled is in
`lsp-on-change` function. Specifically, if the variable is nil, it
calls `(lsp--workspace-sync-method workspace)` which presumably
implies that the description "use method recommended by the
lang-server" is the correct one. So remove the other.

And while at it, remove the `lsp--sync-none` which was never used.

Fixes:
    lsp-mode.el:489:19: Warning: in defcustom for ‘lsp-document-sync-method’:
        duplicated value in ‘choice’: ‘nil’

* lsp-mode.el: swap order of :tag and value in defcustom

Fixes warning:

    lsp-mode.el:2150:25: Warning: in defcustom for ‘lsp-progress-function’:
        misplaced :tag keyword in ‘const’ type

* lsp-mode.el: don't quote condition-case handlers

And while at it, replace an unused condition-case binding to _err with
nil.

Fixes warnings:

    lsp-mode.el:1995:19: Warning: ‘condition-case’ condition should not be quoted:
        'quit

    lsp-mode.el:8686:7: Warning: ‘condition-case’ condition should not be quoted:
        'error

    In lsp--find-root-interactively:
    lsp-mode.el:8990:7: Warning: ‘condition-case’ condition should not be quoted:
        'quit

* lsp-diagnostics.el: declare arguments to a `list` in defcustom :type

Fixes a warning:
    lsp-diagnostics.el:68:10: Warning: in defcustom for
        ‘lsp-diagnostics-attributes’: ‘list’ without arguments