treewide: replace :type list to alist for Action Alist vars
All variables whose types being fixed in this commit are used as an
arg to `(display-buffer-in-side-window)` function. From the docs it
seems that the type of the key and value in this parameter is
completely arbitrary, and since `alist` type-definition allows
omitting the arg we do that.

Fixes a couple of byte-compilation warnings like:

    ‘list’ without arguments