Fixing defcustom warnings + 1 wrong usage of `eq` (#4377)
* treewide: replace :type list to alist for Action Alist vars

All variables whose types being fixed in this commit are used as an
arg to `(display-buffer-in-side-window)` function. From the docs it
seems that the type of the key and value in this parameter is
completely arbitrary, and since `alist` type-definition allows
omitting the arg we do that.

Fixes a couple of byte-compilation warnings like:

    ‘list’ without arguments

* lsp-asm.el: fix two :type definitions

Fixes warnings:
    lsp-asm.el:38:10: Warning: in defcustom for ‘lsp-asm-active-modes’: ‘list’ without arguments
    lsp-ido.el:64:10: Warning: in defcustom for ‘lsp-ido-symbol-kind-to-string’: ‘vector’ without arguments

* lsp-headerline.el: fix wrong usage of (eq) for comparison to ""

`(eq)` does not care about object content, the author probably meant
to use (equal). Either way, there's a special func for string
emptiness, let's use it.

Fixes a warning:
    lsp-headerline.el:400:14: Warning: ‘eq’ called with literal string that may never match (arg 2)