Fix project root discovery when project features aren't loaded. (#4463)
It's possible for the corresponding project package (i.e., projectile
or project) to not be loaded yet but there to exist autoloads for the
functions.  In that situation, using `featurep` to check for the
existence of the package will fail.  Therefore, `fboundp` is used
instead to check for the existence (possibly an autoload) of the
needed functions.