Remove Alire support from lsp-mode Ada client.
Alire support is now available directly in the Ada Language Server and
handles the functionality currently implemented in lsp-mode.  In
addition, when the project file is not specified by the LSP client,
the language server will intelligently locate the project file using
`alr` when an Alire configuration file exists.  Therefore, the default
setting of the project file has been changed (from "default.gpr" to
nil) to enable the server to automatically locate the project file in
normal situations.  The server will also automatically find a project
file for a non-Alire project if it uniquely exists in the root
folder (i.e., no other project file exists in this folder).  The new
project file default setting provides for a better out-of-the-box
configuration to support common situations.