feat: add `lsp-yaml-schema-extensions` (#4527)
* feat: add `lsp-yaml-schema-extensions`

This commit replaces defconst `lsp-yaml--built-in-kubernetes-schema`
with defcustom `lsp-yaml-schema-extensions` to
enable users to extend the dropdown options provided by
`lsp-yaml--get-supported-schemas`. This is helpful since users are no
longer locked into the yaml-language-server's version of the Kubernetes
schema. This also enables users to provide more schemas and multiple
versions of schemas if need be i.e. multiple versions of the Kubernetes
schema, if you work across multiple clusters.

* fix: replace current Kubernetes schema with v1.30.3

The current schema provided `yaml-language-server` does not exist any
longer so is causing a bug when loading the schema.

* fix: docstring for `lsp-yaml-schema-extensions`