Fix logic for rust-analyzer.checkOnSave.features setting inheritance (#4542)
The rust-analyzer inherits the setting of `rust-analyzer.checkOnSve.features`
from `rust-analyzer.cargo.features` by default, but the elisp code here was
incorrectly defaulting the former to the empty list, regardless of the value of
the latter. The effect was that `rust-analyzer.cargo.features` disabled warnings
about "inactive code", without actually turning on type checking of that code!